PythonFastApi/requirements.txt
fastapi[standard]
psycopg[binary,pool]
uvicorn[standard]
python-dotenv
